- Updated references. - Fixed Settings Layout
